разбила тесты на resource и users в разные файлы
